refactor: `UserDatabase`의 메서드 정의를 인터페이스로 추출
다형성을 이용하기 위하여 `UserDatabase`의 메서드를 인터페이스로 추출하고, 구현체의 이름을 `UserMemoryDatabase`로 변경함.
