+++ src/main/java/codesquad/application/config/DatabaseConfig.java
@@ -3,13 +3,14 @@ package codesquad.application.config;
 import codesquad.application.bean.Bean;
 import codesquad.application.database.ArticleDatabase;
 import codesquad.application.database.UserDatabase;
+import codesquad.application.database.UserMemoryDatabase;
 import codesquad.application.database.SessionStorage;
 
 public class DatabaseConfig {
 
     @Bean
     public UserDatabase userDatabase() {
-        return new UserDatabase();
+        return new UserMemoryDatabase();
     }
 
     @Bean
