+++ src/main/java/codesquad/application/database/UserDatabase.java
@@ -1,24 +1,13 @@
 package codesquad.application.database;
 
 import codesquad.application.model.User;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
 
-public class UserDatabase {
-    private final Map<String, User> database = new ConcurrentHashMap<>();
+public interface UserDatabase {
+    void addUser(User user);
 
-    public void addUser(User user) {
-        database.put(user.getUserId(), user);
-    }
+    Optional<User> findUserByUserId(String userId);
 
-    public Optional<User> findUserByUserId(String userId) {
-        return Optional.ofNullable(database.get(userId));
-    }
-
-    public List<User> findAll() {
-        return database.values().stream().sorted(Comparator.comparing(User::getName)).toList();
-    }
+    List<User> findAll();
 }
